add: delete comments, user report

+++ src/dao/commentsDAO.js
@@ -98,13 +98,23 @@ export default class CommentsDAO {
     */
 
     try {
-      // TODO Ticket: Delete Comments
-      // Use the userEmail and commentId to delete the proper comment.
-      const deleteResponse = await comments.deleteOne({
+      const check = await comments.findOne({
         _id: ObjectId(commentId),
       })
 
-      return deleteResponse
+      // console.log("check", check)
+
+      if (check.email === userEmail) {
+        // TODO Ticket: Delete Comments
+        // Use the userEmail and commentId to delete the proper comment.
+        const deleteResponse = await comments.deleteOne({
+          _id: ObjectId(commentId),
+        })
+
+        return deleteResponse
+      } else {
+        return { deletedCount: 0 }
+      }
     } catch (e) {
       console.error(`Unable to delete comment: ${e}`)
       return { error: e }
@@ -122,15 +132,33 @@ export default class CommentsDAO {
     try {
       // TODO Ticket: User Report
       // Return the 20 users who have commented the most on MFlix.
-      const pipeline = []
+      const pipeline = [
+        {
+          $group: {
+            _id: "$email",
+            count: {
+              $sum: 1,
+            },
+          },
+        },
+        {
+          $sort: { count: -1 },
+        },
+        {
+          $limit: 20,
+        },
+      ]
 
       // TODO Ticket: User Report
       // Use a more durable Read Concern here to make sure this data is not stale.
-      const readConcern = comments.readConcern
 
       const aggregateResult = await comments.aggregate(pipeline, {
-        readConcern,
+        readConcern: { level: "majority" },
       })
+      console.log(
+        "CommentsDAO -> mostActiveCommenters -> aggregateResult",
+        aggregateResult,
+      )
 
       return await aggregateResult.toArray()
     } catch (e) {
